task-2-rbac/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==1.4.49
psycopg2-binary==2.9.7
pydantic[email]==1.10.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-multipart==0.0.6
tenacity==8.2.3  